refactor: revise folder layer of left-panel
