+++ packages/core/src/web/app/components/beambox/LeftPanel/components/DrawingToolButtonGroup.spec.tsx
