+++ packages/core/src/web/app/components/beambox/LeftPanel/components/LeftPanelButton.module.scss
