+++ packages/core/src/web/app/components/beambox/LeftPanel/components/LeftPanelSegmented.module.scss
