+++ packages/core/src/web/app/components/beambox/LeftPanel/components/PreviewToolButtonGroup.spec.tsx
