+++ packages/core/src/web/app/components/beambox/LeftPanel/index.module.scss
